Added redux toolkit, a little bit of refactoring
