+++ frontend/src/AppContainer.tsx
@@ -1,4 +1,5 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
+
 import App from './App';
 import { Registration } from './components/Registration';
